# forge-go/api/openapi/codegen-openapi.json
{
  "components": {
    "schemas": {
...
      "Body_uploadFileForGuild": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "file_meta": {
            "nullable": true,
            "title": "File Meta",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadFileForGuild",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Rustic AI API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/guilds/{guild_id}/files/": {
      "get": {
        "operationId": "listFilesForGuild",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Files For Guild"
      },
      "post": {
        "operationId": "uploadFileForGuild",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadFileForGuild"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File For Guild"
      }
    },
...
  }
}